Login and register moved to resp comp
